start-doing/backend/requirements.txt
Flask
Werkzeug
gunicorn
mysql-connector-python
flask-cors
